# src/docs/swagger.yaml
definitions:
...
  models.ParticipantWithUserName:
    properties:
      has_responded:
        type: boolean
      is_host:
        type: boolean
      meeting_id:
        type: integer
      user_name:
        type: string
    type: object
  models.User:
    properties:
      can_login:
        type: boolean
      created_at:
        type: string
      email_address:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      password:
        type: string
      updated_at:
        type: string
      user_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/restricted/participants/meeting/{meeting_id}:
    get:
      consumes:
      - application/json
      description: Get participant by meeting id
      parameters:
      - description: Meeting ID
        in: path
        name: meeting_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ParticipantWithUserName'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get participant by meeting id
      tags:
      - participants
    put:
      consumes:
      - application/json
      description: Update participant by meeting id
      parameters:
      - description: Meeting ID
        in: path
        name: meeting_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ParticipantWithUserName'
        "400":
          description: Error message
          schema:
            type: string
      summary: Update participant by meeting id
      tags:
      - participants
  /api/restricted/participants/new:
    post:
      consumes:
      - application/json
      description: Create participants and other details
      parameters:
      - description: List of ParticipantWithUserName
        in: body
        name: participant
        required: true
        schema:
          $ref: '#/definitions/models.ParticipantWithUserName'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ParticipantWithUserName'
        "400":
          description: Error message
          schema:
            type: string
      summary: Create new participants
      tags:
      - participants
  /api/restricted/signup:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: Details of candidate time
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Error message
          schema:
            type: string
      summary: Create a new user
...
  /user/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: New password
          schema:
            type: string
        "400":
          description: Error message
          schema:
            type: string
      summary: Create a new user
swagger: "2.0"